+++ src/modules/home/components/HomeNavbar/SideMenu/SideMenu.spec.tsx
@@ -1,5 +1,5 @@
 import { shallow } from 'enzyme';
-import SideMenu from 'Components/Layouts/Home/HomeNavbar/SideMenu/index';
+import SideMenu from 'src/modules/home/components/HomeNavbar/SideMenu/index';
 import React from 'react';
 
 describe('Components/SideMenu test', () => {
