+++ src/modules/home/components/Layout/HomeLayout.spec.tsx
@@ -1,4 +1,4 @@
-import HomeLayout from 'Components/Layouts/Home';
+import HomeLayout from 'src/modules/home/components/Layout';
 import { shallow } from 'enzyme';
 import React from 'react';
 import { Layout } from 'antd';
