+++ src/modules/home/pages/styled.tsx
@@ -1,0 +1,33 @@
+import styled, { CSSProperties } from 'styled-components';
+
+export const Icon = styled.div`
+  font-size: 40px;
+`;
+
+export const gridStyle: CSSProperties = {
+  textAlign: 'center',
+};
+
+export const HeaderWrapper = styled.div`
+  width: 100vw;
+  height: 50vh;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  text-align: center;
+
+  background-image: url('https://pbs.twimg.com/media/ElVxJ8OWMAAnew5.jpg');
+  background-size: cover;
+  background-position: center center;
+  background-repeat: no-repeat;
+  background-attachment: fixed;
+`;
+
+export const Heading = styled.div`
+  font-size: 5em;
+  color: #ffffff;
+  margin-top: 0;
+  margin-bottom: 0em;
+`;
