+++ src/pages/_app.tsx
@@ -1,11 +1,10 @@
-import 'antd/dist/antd.css';
-import HomeNavbar from 'Components/Layouts/Home/HomeNavbar';
+import HomeNavbar from 'Modules/home/components/HomeNavbar';
 import { AppProps } from 'next/dist/next-server/lib/router/router';
 import { useRouter } from 'next/router';
 import React from 'react';
 import { Provider } from 'react-redux';
-import { store } from 'src/common/stores/stores';
-import { GlobalStyle } from 'src/common/styles/Global.styled';
+import { store } from 'Stores/stores';
+import { GlobalStyle } from 'Styles/Global.styled';
 
 const EXCLUDE_NAVBAR: string[] = ['/login'];
